fix daily deck label on login
